+++ pybop/optimisers/base_pints_optimiser.py
@@ -198,13 +198,7 @@ class BasePintsOptimiser(BaseOptimiser):
 
         # Choose method to evaluate
         def fun(x):
-            if self._needs_sensitivities:
-                L, dl = self.cost(x, calculate_grad=True, apply_transform=True)
-            else:
-                L = self.cost(x, apply_transform=True)
-                dl = None
-            sign = -1 if not self.minimising else 1
-            return (sign * L, sign * dl) if dl is not None else sign * L
+            return self.cost_call(x, calculate_grad=self._needs_sensitivities)
 
         # Create evaluator object
         if self._parallel:
@@ -221,9 +215,6 @@ class BasePintsOptimiser(BaseOptimiser):
 
         # Keep track of current best and best-guess scores.
         fb = fg = np.inf
-
-        # Internally we always minimise! Keep a 2nd value to show the user.
-        fg_user = (fb, fg) if self.minimising else (-fb, -fg)
 
         # Keep track of the last significant change
         f_sig = np.inf
@@ -244,7 +235,6 @@ class BasePintsOptimiser(BaseOptimiser):
                 # Update the scores
                 fb = self.optimiser.f_best()
                 fg = self.optimiser.f_guessed()
-                fg_user = (fb, fg) if self.minimising else (-fb, -fg)
 
                 # Check for significant changes against the absolute and relative tolerance
                 f_new = fg if self._use_f_guessed else fb
@@ -263,8 +253,8 @@ class BasePintsOptimiser(BaseOptimiser):
                 self.log_update(
                     x=xs,
                     x_best=self.optimiser.x_best(),
-                    cost=_fs if self.minimising else [-x for x in _fs],
-                    cost_best=fb if self.minimising else -fb,
+                    cost=_fs,
+                    cost_best=fb,
                 )
 
                 # Check stopping criteria:
@@ -327,14 +317,11 @@ class BasePintsOptimiser(BaseOptimiser):
             # Show last result and exit
             print("\n" + "-" * 40)
             print("Unexpected termination.")
-            print("Current score: " + str(fg_user))
+            print("Current score: " + str((fb, fg)))
             print("Current position:")
 
-            # Show current parameters
-            x_user = self.optimiser.x_guessed()
-            if self._transformation:
-                x_user = self._transformation.to_model(x_user)
-            for p in x_user:
+            # Show current parameters (with any transformation applied)
+            for p in self.optimiser.x_guessed():
                 print(PintsStrFloat(p))
             print("-" * 40)
             raise
@@ -355,16 +342,11 @@ class BasePintsOptimiser(BaseOptimiser):
             x = self.optimiser.x_best()
             f = self.optimiser.f_best()
 
-        # Inverse transform search parameters
-        if self._transformation:
-            x = self._transformation.to_model(x)
-
         return OptimisationResult(
-            x=x,
-            cost=self.cost,
-            final_cost=f if self.minimising else -f,
-            n_iterations=self._iterations,
             optim=self,
+            x=x,
+            final_cost=f,
+            n_iterations=self._iterations,
             time=total_time,
         )